requirements.txt
numpy == 1.14.5
matplotlib == 2.2.2
gdal == 2.2.2
lcmap-merlin == 2.2.0
PyQt5 >= 5.6
pyyaml
cytoolz
cython
requests
